+++ Communication/Packets/Incoming/Marketplace/GetOwnOffersEvent.cs
@@ -1,17 +1,16 @@
-﻿using Plus.Communication.Packets.Outgoing.Marketplace;
+﻿using Plus.HabboHotel.Catalog.Marketplace;
 using Plus.HabboHotel.GameClients;
 
 namespace Plus.Communication.Packets.Incoming.Marketplace;
 
 internal class GetOwnOffersEvent : IPacketEvent
 {
-    public Task Parse(GameClient session, IIncomingPacket packet)
-    {
-        var habbo = session.GetHabbo();
-        if (habbo == null)
-            return Task.CompletedTask;
+    private readonly IMarketplaceService _marketplaceService;
 
-        session.Send(new MarketPlaceOwnOffersComposer(habbo.Id));
-        return Task.CompletedTask;
+    public GetOwnOffersEvent(IMarketplaceService marketplaceService)
+    {
+        _marketplaceService = marketplaceService;
     }
+
+    public Task Parse(GameClient session, IIncomingPacket packet) => _marketplaceService.GetOwnOffers(session);
 }